+++ cheetah/accelerator/custom_transfer_map.py
@@ -62,14 +62,27 @@ class CustomTransferMap(Element):
             " incorrect tracking results."
         )
 
-        device = elements[0].transfer_map(incoming_beam.energy).device
-        dtype = elements[0].transfer_map(incoming_beam.energy).dtype
+        device = (
+            elements[0]
+            .transfer_map(incoming_beam.energy, incoming_beam.species.mass_eV)
+            .device
+        )
+        dtype = (
+            elements[0]
+            .transfer_map(incoming_beam.energy, incoming_beam.species.mass_eV)
+            .dtype
+        )
 
         tm = torch.eye(7, device=device, dtype=dtype).repeat(
             (*incoming_beam.energy.shape, 1, 1)
         )
         for element in elements:
-            tm = torch.matmul(element.transfer_map(incoming_beam.energy), tm)
+            tm = torch.matmul(
+                element.transfer_map(
+                    incoming_beam.energy, incoming_beam.species.mass_eV
+                ),
+                tm,
+            )
             incoming_beam = element.track(incoming_beam)
 
         combined_length = sum(element.length for element in elements)
@@ -80,7 +93,9 @@ class CustomTransferMap(Element):
             tm, length=combined_length, device=device, dtype=dtype, name=combined_name
         )
 
-    def transfer_map(self, energy: torch.Tensor) -> torch.Tensor:
+    def transfer_map(
+        self, energy: torch.Tensor, particle_mass_eV: torch.Tensor
+    ) -> torch.Tensor:
         return self.predefined_transfer_map
 
     @property