+++ cheetah/accelerator/element.py
@@ -25,7 +25,9 @@ class Element(ABC, nn.Module):
         self.name = name if name is not None else generate_unique_name()
         self.register_buffer("length", torch.tensor(0.0, device=device, dtype=dtype))
 
-    def transfer_map(self, energy: torch.Tensor) -> torch.Tensor:
+    def transfer_map(
+        self, energy: torch.Tensor, particle_mass_eV: torch.Tensor
+    ) -> torch.Tensor:
         r"""
         Generates the element's transfer map that describes how the beam and its
         particles are transformed when traveling through the element.
@@ -49,6 +51,7 @@ class Element(ABC, nn.Module):
         affine transformation.
 
         :param energy: Reference energy of the Beam. Read from the fed-in Cheetah Beam.
+        :param particle_mass_eV: Mass of the beam's particle species in eV.
         :return: A 7x7 Matrix for further calculations.
         """
         raise NotImplementedError
@@ -62,7 +65,7 @@ class Element(ABC, nn.Module):
         :return: Beam of particles exiting the element.
         """
         if isinstance(incoming, ParameterBeam):
-            tm = self.transfer_map(incoming.energy)
+            tm = self.transfer_map(incoming.energy, incoming.species.mass_eV)
             mu = torch.matmul(tm, incoming._mu.unsqueeze(-1)).squeeze(-1)
             cov = torch.matmul(tm, torch.matmul(incoming._cov, tm.transpose(-2, -1)))
             return ParameterBeam(
@@ -74,7 +77,7 @@ class Element(ABC, nn.Module):
                 dtype=mu.dtype,
             )
         elif isinstance(incoming, ParticleBeam):
-            tm = self.transfer_map(incoming.energy)
+            tm = self.transfer_map(incoming.energy, incoming.species.mass_eV)
             new_particles = torch.matmul(incoming.particles, tm.transpose(-2, -1))
             return ParticleBeam(
                 new_particles,